requirements.txt
fastapi
uvicorn
scikit-learn
numpy
scipy
groq
python-dotenv
sentence-transformers
python-telegram-bot
httpx
pytest

